bug-fixing and unittest issue

+++ headers/Graph.hpp
@@ -16,12 +16,12 @@ public:
     std::vector<Vertice*> m_vertices;
 
     //constructors:
-    Graph(size_t numberOfNodes): m_numberOfNodes(numberOfNodes)
+    Graph(int numberOfNodes): m_numberOfNodes(numberOfNodes)
     {
         InitializeVerticeVector();
     };
 
-    Graph(size_t numberOfNodes, bool isDirected): m_numberOfNodes(numberOfNodes), m_isDirected(isDirected)
+    Graph(int numberOfNodes, bool isDirected): m_numberOfNodes(numberOfNodes), m_isDirected(isDirected)
     {
         InitializeVerticeVector();
     };
@@ -29,7 +29,7 @@ public:
     void AddEdge(Vertice*, Vertice*);
     void SortVertices(void);
 private:
-    size_t  m_numberOfNodes;
+    int  m_numberOfNodes;
     void InitializeVerticeVector();
 };
